made some small changes

+++ Lab2/Task2_NC/Task2.c
@@ -1,16 +1,17 @@
 #include "Task2.h"
 
-void bit_set(int location, int16_t sel) {
+void bit_set(int location, int8_t sel) {
     int16_t bit = (1 << sel);
     location |= bit;
 }
 
-void bit_clr(int location, int16_t sel) {
+void bit_clr(int location, int8_t sel) {
     int16_t bit = (1 << sel);
     location &= !bit;
 }
 
-void bit_tog(int location, int16_t sel) {
+void bit_tog(int location, int8_t sel) {
     int16_t bit = (1 << sel);
     location ^= bit;
 }
+
